Fix errors with Trace level enabled.
When enabling Trace level there is a lot of noise. Unfortunately because
of `Arbiter::spawn` there were log messages sent, but in the same time
when the log rotation didn't finish yet there was a Trace message going
from inside of actix-web therefore causing a panic and a PoisonError on
mutexes.