Update to use post strategy
The streaming strategy had several problems, as did the C bindings
to libz4 this replaces that with the pure rust version of libflate
which is quite fast and performs well enough on small buffers that
streaming compression isn't needed.

The lack of streaming compression lets us simplify the code some and
remove the hard dependency on actix and replace it with a simlpe dependency
on futures and a futures supporting web client. Both of which will be in std
in the near future