Upgrade for async/await
This upgrades compressed log for async/await. This upgrade has some
actual execution time differences that need to be accounted for.

Namely that instead of using the already running actix arbiter
(a very specific thing that made this crate not very useful to
those not integrated into the actix ecosystem) it spawns an actix
exectutor. Runs the one upload and then cleans it up.

I don't actually know what the overhead of this process is. I don't
think it's too bad. But it's worth considering.

The other semantic change is that 'flush' now actually performs it's
function by flushing the buffer and rotating it. I'm not sure if this
is really the best idea. As the flush operation is designed for disks
and may be called more often than is justified when our goal is to
increase compression.

That being said, perhaps the calling function has a good reason that
we should flush now. An important error has occured or any other reason.

The final most minor change is the removal of the failure crate.