Spawn Actix executor for each upload event
While the last version of this crate 'worked' it really pushed those
quotation marks. Log uploads would only actually occur when you happened
to have an actix system lying around in the same thread. Somthing that's
not garunteed, especially since the logger is shared between all threads
and whatever thread pushes it over the limit is where the upload
actually happens. This patch fixes the issue and at the expense of some
additional overhead every upload will actually ensure that there is an
Actix system around to do the job.