Upgrade to actix 11 and awc 3.0
This ptach upgrades the actix client creates to match the rest of
Althea's rust suite. Testing shows that an executor problem of some kind
was preventing proper logging, with this change things seem to bein the
right again. I wish I had a better idea of how exactly that occured,
silent upgrades to packges causing breakage like this is a scary
makes you hesitate to update your Cargo.lock.
It's possibility but I'm not totally sure that's what's happening.