Do upload in new thread
This spawns the upload operation in a new thread, this is required
because the new upload code tries to adopt the executor of the current
thread rather than the previous design, that required an active actix
executor in the present thread in order to make any progress.

This new design is more resource intensive, but occurs pretty rarely as
logs are only uploaded every few minutes. It's also far more flexible.